requirements.txt
Django>=3.2,<4.0
PyYAML>=6.0
psycopg2-binary>=2.9
yfinance>=0.2.18
pandas>=1.5.3
matplotlib>=3.6.3
mplfinance>=0.12.9b7
